fix: pool share and types

+++ contracts/src/interface/IDexterity.sol
@@ -25,5 +25,5 @@ interface IDexterity {
   function createPool(address firstToken, address secondToken) external;
   function getPool(address firstToken, address secondToken) external view returns (Pool memory);
   function deposit(address firstToken, address secondToken, uint128 firstAmount, uint128 secondAmount) external;
-  function withdraw(address firstToken, address secondToken, uint256 shares) external;
+  function withdraw(address firstToken, address secondToken, uint128 shares) external;
 }
